Trickling_up

Documentation for Trickling_up.

Trickling_up.simulate_baseline

Trickling_up.simulate_monetary

Trickling_up.simulate_re

Trickling_up.solver

Trickling_up.simulate_baseline — Method

Simulates a baseline model

Trickling_up.simulate_monetary — Method

Extends the model to include a monetary policy response.

Trickling_up.simulate_re — Method

Extends the model to include rational expectations of future income

Trickling_up.solver — Method

(Very) lightweight solver for linear rational expectations models in continuous time

Suppose our model is given by the linear ODE

[xdot] = [Axx Axy] [x] [ydot] [Ayx Ayy] [y]

where x is a predetermined 'state' vector of length N, y is vector of 'jump' variables, and the four blocks A_ form a matrix A. For there to be a unique solution for y given the state x, there must be N stable eigenvalues (i.e. with negative real part) of A.

Let A = QUQ' be the Schur decomposition, where Q is unitary and U is upper triangular. Assuming that this is sorted so that the stable eigenvalues of U are at the top left, we can write.

U = [Uss Usu] Q = [Qxs Qxu] [0 Uuu] [Qys Q_yu]

where the N-by-N upper left block Uss maps the stable subspace to itself, the N-by-N Qxs maps the stable subspace to x, etc.

Given any state x, we can obtain the rotated stable state by solving s = Qxs^(-1)*x. Then sdot = Usss, and finally xdot = Q_xssdot = Qxs*UssQxs^(-1). The jump variables y can be obtained from s by y = Qyxs = Qyx*Qxs^(-1)*x.

Summing up, our model will have law of motion

xdot = Bx y = Fx

where B = Qxs*UssQxs^(-1) and F = QyxQ_xs^(-1).